HW-1:Построена иерархия животных. А также 3 метода по созданию уникальных животных.
