--- HW-1/src/main/java/AbstractAnimal.java
@@ -1,23 +1,0 @@
-public class AbstractAnimal implements Animal{
-    protected String breed;
-    protected String name;
-    protected Double cost;
-    protected String character;
-
-    @Override
-    public String getBreed() {
-        return breed;
-    }
-    @Override
-    public String getName() {
-        return name;
-    }
-    @Override
-    public Double getCost() {
-        return cost;
-    }
-    @Override
-    public String getCharacter() {
-        return character;
-    }
-}
