--- HW-1/src/main/java/Animal.java
@@ -1,6 +1,0 @@
-public interface Animal {
-    String getBreed();
-    String getName();
-    Double getCost();
-    String getCharacter();
-}
