+++ HW-1/src/main/java/AnimalGeneral/Animal.java
@@ -1,0 +1,23 @@
+package AnimalGeneral;
+
+/**
+ * Интерфейс, описывающий поведение животных
+ */
+public interface Animal {
+    /**
+     * Метод для получения породы
+     */
+    String getBreed();
+    /**
+     * Метод для получения имени
+     */
+    String getName();
+    /**
+     * Метод для получения стоимости
+     */
+    Double getCost();
+    /**
+     * Метод для получения характера
+     */
+    String getCharacter();
+}
