+++ HW-1/src/main/java/AnimalGeneral/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Пакет, содержащий общий функционал для работы с животными
+ */
+package AnimalGeneral;
