+++ HW-1/src/main/java/AnimalsClassifiers/Pet.java
@@ -1,0 +1,10 @@
+package AnimalsClassifiers;
+
+import AnimalGeneral.AbstractAnimal;
+
+/**
+ * Это родительский класс для всех домашних животных
+ */
+public abstract class Pet extends AbstractAnimal {
+
+}
