+++ HW-1/src/main/java/AnimalsClassifiers/Predator.java
@@ -1,0 +1,9 @@
+package AnimalsClassifiers;
+
+import AnimalGeneral.AbstractAnimal;
+
+/**
+ * Это родительский класс для всех хищных животных
+ */
+public abstract class Predator extends AbstractAnimal {
+}
