+++ HW-1/src/main/java/AnimalsClassifiers/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Пакет, содержащий классификаторы животных
+ */
+package AnimalsClassifiers;
