--- HW-1/src/main/java/Cat.java
@@ -1,2 +1,0 @@
-public class Cat extends Pet{
-}
