--- HW-1/src/main/java/Cow.java
@@ -1,2 +1,0 @@
-public class Cow extends Pet{
-}
