--- HW-1/src/main/java/Dog.java
@@ -1,2 +1,0 @@
-public class Dog extends Pet{
-}
