--- HW-1/src/main/java/Lion.java
@@ -1,2 +1,0 @@
-public class Lion extends Predator{
-}
