+++ HW-1/src/main/java/Main.java
@@ -1,4 +1,14 @@
+import AnimalGeneral.CreateAnimalService;
+import AnimalGeneral.CreateAnimalServiceImpl;
+
+/**
+ * Главный класс с вызовом метода main
+ */
 public class Main {
+    /**
+     * Точка старта приложения
+     * @param args - входящие параметры
+     */
     public static void main(String[] args) {
         CreateAnimalService createAnimalService = new CreateAnimalService();
         System.out.println("\nWhile cycle:");
