--- HW-1/src/main/java/Pet.java
@@ -1,3 +1,0 @@
-public class Pet extends AbstractAnimal{
-
-}
