+++ HW-1/src/main/java/PetAnimals/Cat.java
@@ -1,0 +1,9 @@
+package PetAnimals;
+
+import AnimalsClassifiers.Pet;
+
+/**
+ * Класс для создания кота
+ */
+public class Cat extends Pet {
+}
