+++ HW-1/src/main/java/PetAnimals/Cow.java
@@ -1,0 +1,9 @@
+package PetAnimals;
+
+import AnimalsClassifiers.Pet;
+
+/**
+ * Класс для создания коровы
+ */
+public class Cow extends Pet {
+}
