+++ HW-1/src/main/java/PetAnimals/Dog.java
@@ -1,0 +1,9 @@
+package PetAnimals;
+
+import AnimalsClassifiers.Pet;
+
+/**
+ * Класс для создания собаки
+ */
+public class Dog extends Pet {
+}
