+++ HW-1/src/main/java/PetAnimals/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Пакет, содержащий классы домашних животных
+ */
+package PetAnimals;
