--- HW-1/src/main/java/Predator.java
@@ -1,2 +1,0 @@
-public class Predator extends AbstractAnimal {
-}
