+++ HW-1/src/main/java/PredatorAnimals/Lion.java
@@ -1,0 +1,9 @@
+package PredatorAnimals;
+
+import AnimalsClassifiers.Predator;
+
+/**
+ * Класс для создания льва
+ */
+public class Lion extends Predator {
+}
