+++ HW-1/src/main/java/PredatorAnimals/Shark.java
@@ -1,0 +1,9 @@
+package PredatorAnimals;
+
+import AnimalsClassifiers.Predator;
+
+/**
+ * Класс для создания акулы
+ */
+public class Shark extends Predator {
+}
