+++ HW-1/src/main/java/PredatorAnimals/Wolf.java
@@ -1,0 +1,9 @@
+package PredatorAnimals;
+
+import AnimalsClassifiers.Predator;
+
+/**
+ * Класс для создания волка
+ */
+public class Wolf extends Predator {
+}
