+++ HW-1/src/main/java/PredatorAnimals/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Пакет, содержащий классы хищных животных
+ */
+package PredatorAnimals;
