--- HW-1/src/main/java/Shark.java
@@ -1,2 +1,0 @@
-public class Shark extends Predator{
-}
