--- HW-1/src/main/java/Wolf.java
@@ -1,2 +1,0 @@
-public class Wolf extends Predator{
-}
